Adds -ghc-version flag to ghc.
Summary:
When building the rts with ghc (e.g. using ghc as a c compiler), ghc's
Added Value includes adding `-include /path/to/ghcversion.h`.  For
this it looksup the rts package in the package database, which--if
empty--fails.  Thus to allow compiling C files with GHC, we add the
`-ghc-version` flag, which takes the path to the `ghcversion.h` file.

A `-no-ghc-version` flag was omitted, as at that point it becomes
questionable why one would use ghc to compile c if one doesn't
any of the added value.

Reviewers: bgamari, geekosaur, austin

Subscribers: rwbarton, thomie

Differential Revision: https://phabricator.haskell.org/D4135