Adds rts/rts.cabal.in file
Summary:
This is in prerpation for cabalification of the `rts`. To be actually able to parse this file, a rather recent Cabal is required. One after commit `357d49d` of haskell/cabal. The relevant PR to support the new `asm-srouces` and `cmm-sources` is haskell/cabal/pull/4857.

Not that this does *not* allow cabal to build the RTS. It does however provide enough information such that cabal can `copy` and `register` the package properly in the package database, if all the build artifacts have been build properly.

As such it does not require any custom handling of the `rts` package. As the rts as well as all the other packages built by the GHC built system are built outside of cabal anyway.

Reviewers: bgamari, hvr, erikd, simonmar

Subscribers: rwbarton, thomie, erikd

Differential Revision: https://phabricator.haskell.org/D4174