Use LICENSE instead of ../LICENSE in the compiler.cabal file
Summary: The `compiler` directory contains a `LICENSE` file, which is identical to the one in `../LICENSE`. As such we can just use the LICENSE file that accompanies the `compiler` instead of tyring to fish it out from somewhere higher up in the source tree.  This also makes the compiler package more self contained.

Reviewers: bgamari

Subscribers: rwbarton, thomie

Differential Revision: https://phabricator.haskell.org/D4175