Allow the rts lib to be called rts-1.0
Summary: When using the `rts.cabal` file the rts lirbary ends up being `rts-1.0` instead of `rts`. As sucht it's called `libHSrts-1.0` instead of `libHSrts`. A fun consequence of this is, that when asking ghc to link a `-treaded` program, it fails to mutate the `rts` lib selection into the threaded variant.

Reviewers: bgamari, simonmar

Subscribers: rwbarton, thomie

Differential Revision: https://phabricator.haskell.org/D4187

# Conflicts:
#	hadrian